requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
numpy=1.20.1
python=3.8.8
pytorch=1.8.1
scipy=1.6.2
torchvision=0.9.1
